backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
requests==2.31.0
pandas==2.1.3
numpy==1.25.2
pydantic==2.5.0
python-multipart==0.0.6
aiohttp==3.9.0
python-dotenv==1.0.0
redis==5.0.1
schedule==1.2.0
plotly==5.17.0
anthropic>=0.40.0
typing-extensions==4.8.0
python-dateutil==2.8.2
pytz==2023.3
